Latest changes with all cars setup

+++ siteParent/site_WEB/src/main/java/site/racinggame/RaceResult.java
@@ -1,5 +1,7 @@
 package site.racinggame;
 
+import java.math.BigDecimal;
+
 
 public class RaceResult implements java.io.Serializable{
 	
@@ -8,17 +10,17 @@ public class RaceResult implements java.io.Serializable{
 	private String racingClass;
 	private String raceType;
 	private String place1;
-	private Double place1Time;
+	private BigDecimal place1Time;
 	private String place2;
-	private Double place2Time;
+	private BigDecimal place2Time;
 	private String place3;
-	private Double place3Time;
+	private BigDecimal place3Time;
 	private String place4;
-	private Double place4Time;
+	private BigDecimal place4Time;
 	private String place5;
-	private Double place5Time;
+	private BigDecimal place5Time;
 	private String place6;
-	private Double place6Time;
+	private BigDecimal place6Time;
 	
 	public RaceResult(){
 	}
@@ -47,11 +49,11 @@ public class RaceResult implements java.io.Serializable{
 		this.place1 = place1;
 	}
 
-	public Double getPlace1Time() {
+	public BigDecimal getPlace1Time() {
 		return place1Time;
 	}
 
-	public void setPlace1Time(Double place1Time) {
+	public void setPlace1Time(BigDecimal place1Time) {
 		this.place1Time = place1Time;
 	}
 
@@ -63,11 +65,11 @@ public class RaceResult implements java.io.Serializable{
 		this.place2 = place2;
 	}
 
-	public Double getPlace2Time() {
+	public BigDecimal getPlace2Time() {
 		return place2Time;
 	}
 
-	public void setPlace2Time(Double place2Time) {
+	public void setPlace2Time(BigDecimal place2Time) {
 		this.place2Time = place2Time;
 	}
 
@@ -79,11 +81,11 @@ public class RaceResult implements java.io.Serializable{
 		this.place3 = place3;
 	}
 
-	public Double getPlace3Time() {
+	public BigDecimal getPlace3Time() {
 		return place3Time;
 	}
 
-	public void setPlace3Time(Double place3Time) {
+	public void setPlace3Time(BigDecimal place3Time) {
 		this.place3Time = place3Time;
 	}
 
@@ -95,11 +97,11 @@ public class RaceResult implements java.io.Serializable{
 		this.place4 = place4;
 	}
 
-	public Double getPlace4Time() {
+	public BigDecimal getPlace4Time() {
 		return place4Time;
 	}
 
-	public void setPlace4Time(Double place4Time) {
+	public void setPlace4Time(BigDecimal place4Time) {
 		this.place4Time = place4Time;
 	}
 
@@ -111,11 +113,11 @@ public class RaceResult implements java.io.Serializable{
 		this.place5 = place5;
 	}
 
-	public Double getPlace5Time() {
+	public BigDecimal getPlace5Time() {
 		return place5Time;
 	}
 
-	public void setPlace5Time(Double place5Time) {
+	public void setPlace5Time(BigDecimal place5Time) {
 		this.place5Time = place5Time;
 	}
 
@@ -127,11 +129,11 @@ public class RaceResult implements java.io.Serializable{
 		this.place6 = place6;
 	}
 
-	public Double getPlace6Time() {
+	public BigDecimal getPlace6Time() {
 		return place6Time;
 	}
 
-	public void setPlace6Time(Double place6Time) {
+	public void setPlace6Time(BigDecimal place6Time) {
 		this.place6Time = place6Time;
 	}
